consolidate parsing of descriptions and packages in db.
The strategy for now would be for anything that isn't always required
for the db to be optional. This way we can add new fields to the Package struct.

One consideration is we will need to ultimately balance when we add new fields to the struct, this is what is stored in the packagedb so
adding a new field will break the deseralization at the moment.